+++ docs/release-notes.md
@@ -11,7 +11,7 @@
     * Fixed `--no-terminal` still opening a terminal window, even if nothing is logged to it (thanks to Ryhon0!).
 
 * For mod authors:
-  * Added [`AssetRequested` content event](https://stardewvalleywiki.com/Modding:Modder_Guide/APIs/Events#Content), which will eventually replace `IAssetEditor` and `IAssetLoader`.
+  * Added [content events](https://stardewvalleywiki.com/Modding:Modder_Guide/APIs/Events#Content), which will replace `IAssetEditor` and `IAssetLoader` in SMAPI 4.0.0.
   * Overhauled [mod-provided APIs](https://stardewvalleywiki.com/Modding:Modder_Guide/APIs/Integrations#Mod-provided_APIs) (thanks to Shockah!).  
     _This adds support for many previously-unsupported cases: proxied interfaces in return values or input arguments, proxied enums if their values match, generic methods, and more._
   * Added `Constants.ContentPath`.